requirements.txt
python-dateutil
python-docx

